move tests out of process event - they don't need any of the setup there
