+++ plugin-server/tests/main/ingestion-queues/session-recording/process-event.tests.ts
@@ -1,0 +1,460 @@
+import { DateTime } from 'luxon'
+
+import {
+    ConsoleLogEntry,
+    createPerformanceEvent,
+    createSessionReplayEvent,
+    gatherConsoleLogEvents,
+    getTimestampsFrom,
+    SummarizedSessionRecordingEvent,
+} from '../../../../src/main/ingestion-queues/session-recording/process-events'
+import { RRWebEvent, TimestampFormat } from '../../../../src/types'
+import { castTimestampToClickhouseFormat } from '../../../../src/utils/utils'
+
+describe('session recording process event', () => {
+    const sessionReplayEventTestCases: {
+        snapshotData: { events_summary: RRWebEvent[] }
+        snapshotSource?: string
+        expected: Pick<
+            SummarizedSessionRecordingEvent,
+            | 'click_count'
+            | 'keypress_count'
+            | 'mouse_activity_count'
+            | 'first_url'
+            | 'first_timestamp'
+            | 'last_timestamp'
+            | 'active_milliseconds'
+            | 'console_log_count'
+            | 'console_warn_count'
+            | 'console_error_count'
+            | 'size'
+            | 'event_count'
+            | 'message_count'
+            | 'snapshot_source'
+        >
+    }[] = [
+        {
+            snapshotData: {
+                events_summary: [{ timestamp: 1682449093469, type: 3, data: { source: 2 }, windowId: '1' }],
+            },
+            expected: {
+                click_count: 1,
+                keypress_count: 0,
+                mouse_activity_count: 1,
+                first_url: null,
+                first_timestamp: '2023-04-25 18:58:13.469',
+                last_timestamp: '2023-04-25 18:58:13.469',
+                active_milliseconds: 1, //  one event, but it's active, so active time is 1ms not 0
+                console_log_count: 0,
+                console_warn_count: 0,
+                console_error_count: 0,
+                size: 73,
+                event_count: 1,
+                message_count: 1,
+                snapshot_source: 'web',
+            },
+        },
+        {
+            snapshotData: {
+                events_summary: [{ timestamp: 1682449093469, type: 3, data: { source: 5 }, windowId: '1' }],
+            },
+            expected: {
+                click_count: 0,
+                keypress_count: 1,
+                mouse_activity_count: 1,
+                first_url: null,
+                first_timestamp: '2023-04-25 18:58:13.469',
+                last_timestamp: '2023-04-25 18:58:13.469',
+                active_milliseconds: 1, //  one event, but it's active, so active time is 1ms not 0
+                console_log_count: 0,
+                console_warn_count: 0,
+                console_error_count: 0,
+                size: 73,
+                event_count: 1,
+                message_count: 1,
+                snapshot_source: 'web',
+            },
+        },
+        {
+            snapshotData: {
+                events_summary: [
+                    { timestamp: 1682449093469, type: 3, data: { source: 5 }, windowId: '1' },
+                    {
+                        type: 6,
+                        data: { plugin: 'rrweb/console@1', payload: { level: 'log' } },
+                        timestamp: 1682449093469,
+                        windowId: '1',
+                    },
+                    {
+                        type: 6,
+                        data: { plugin: 'rrweb/console@1', payload: { level: 'log' } },
+                        timestamp: 1682449093469,
+                        windowId: '1',
+                    },
+                    {
+                        type: 6,
+                        data: { plugin: 'rrweb/console@1', payload: { level: 'warn' } },
+                        timestamp: 1682449093469,
+                        windowId: '1',
+                    },
+                    {
+                        type: 6,
+                        data: { plugin: 'rrweb/console@1', payload: { level: 'warn' } },
+                        timestamp: 1682449093469,
+                        windowId: '1',
+                    },
+                    {
+                        type: 6,
+                        data: { plugin: 'rrweb/console@1', payload: { level: 'warn' } },
+                        timestamp: 1682449093469,
+                        windowId: '1',
+                    },
+                    {
+                        type: 6,
+                        data: { plugin: 'rrweb/console@1', payload: { level: 'error' } },
+                        timestamp: 1682449093469,
+                        windowId: '1',
+                    },
+                ],
+            },
+            expected: {
+                click_count: 0,
+                keypress_count: 1,
+                mouse_activity_count: 1,
+                first_url: null,
+                first_timestamp: '2023-04-25 18:58:13.469',
+                last_timestamp: '2023-04-25 18:58:13.469',
+                active_milliseconds: 1, //  one event, but it's active, so active time is 1ms not 0
+                console_log_count: 2,
+                console_warn_count: 3,
+                console_error_count: 1,
+                size: 762,
+                event_count: 7,
+                message_count: 1,
+                snapshot_source: 'web',
+            },
+        },
+        {
+            snapshotData: {
+                events_summary: [
+                    {
+                        timestamp: 1682449093693,
+                        type: 5,
+                        data: {
+                            payload: {
+                                // doesn't match because href is nested in payload
+                                href: 'http://127.0.0.1:8000/home',
+                            },
+                        },
+                        windowId: '1',
+                    },
+                    {
+                        timestamp: 1682449093469,
+                        type: 4,
+                        data: {
+                            href: 'http://127.0.0.1:8000/second/url',
+                        },
+                        windowId: '1',
+                    },
+                ],
+            },
+            expected: {
+                click_count: 0,
+                keypress_count: 0,
+                mouse_activity_count: 0,
+                first_url: 'http://127.0.0.1:8000/second/url',
+                first_timestamp: '2023-04-25 18:58:13.469',
+                last_timestamp: '2023-04-25 18:58:13.693',
+                active_milliseconds: 0, // no data.source, so no activity
+                console_log_count: 0,
+                console_warn_count: 0,
+                console_error_count: 0,
+                size: 213,
+                event_count: 2,
+                message_count: 1,
+                snapshot_source: 'web',
+            },
+        },
+        {
+            snapshotData: {
+                events_summary: [
+                    // a negative timestamp is ignored
+                    { timestamp: 1682449093000, type: 3, data: { source: 2 }, windowId: '1' },
+                    { timestamp: 1682449095000, type: 3, data: { source: 2 }, windowId: '1' },
+                    { timestamp: -922167545571, type: 3, data: { source: 2 }, windowId: '1' },
+                ],
+            },
+            expected: {
+                click_count: 3,
+                keypress_count: 0,
+                mouse_activity_count: 3,
+                first_url: null,
+                first_timestamp: '2023-04-25 18:58:13.000',
+                last_timestamp: '2023-04-25 18:58:15.000',
+                active_milliseconds: 1,
+                console_log_count: 0,
+                console_warn_count: 0,
+                console_error_count: 0,
+                size: 217,
+                event_count: 3,
+                message_count: 1,
+                snapshot_source: 'web',
+            },
+        },
+        {
+            snapshotData: {
+                events_summary: [
+                    // three windows with 1 second, 2 seconds, and 3 seconds of activity
+                    // even though they overlap they should be summed separately
+                    { timestamp: 1682449093000, type: 3, data: { source: 2 }, windowId: '1' },
+                    { timestamp: 1682449094000, type: 3, data: { source: 2 }, windowId: '1' },
+                    { timestamp: 1682449095000, type: 3, data: { source: 2 }, windowId: '2' },
+                    { timestamp: 1682449097000, type: 3, data: { source: 2 }, windowId: '2' },
+                    { timestamp: 1682449096000, type: 3, data: { source: 2 }, windowId: '3' },
+                    { timestamp: 1682449099000, type: 3, data: { source: 2 }, windowId: '3' },
+                ],
+            },
+            expected: {
+                click_count: 6,
+                keypress_count: 0,
+                mouse_activity_count: 6,
+                first_url: null,
+                first_timestamp: '2023-04-25 18:58:13.000',
+                last_timestamp: '2023-04-25 18:58:19.000',
+                active_milliseconds: 6000, // can sum up the activity across windows
+                console_log_count: 0,
+                console_warn_count: 0,
+                console_error_count: 0,
+                size: 433,
+                event_count: 6,
+                message_count: 1,
+                snapshot_source: 'web',
+            },
+        },
+        {
+            snapshotData: {
+                events_summary: [{ timestamp: 1682449093000, type: 3, data: { source: 2 }, windowId: '1' }],
+            },
+            snapshotSource: 'mobile',
+            expected: {
+                active_milliseconds: 1,
+                click_count: 1,
+                console_error_count: 0,
+                console_log_count: 0,
+                console_warn_count: 0,
+                event_count: 1,
+                first_timestamp: '2023-04-25 18:58:13.000',
+                first_url: null,
+                keypress_count: 0,
+                last_timestamp: '2023-04-25 18:58:13.000',
+                message_count: 1,
+                mouse_activity_count: 1,
+                size: 73,
+                snapshot_source: 'mobile',
+            },
+        },
+    ]
+    sessionReplayEventTestCases.forEach(({ snapshotData, snapshotSource, expected }) => {
+        test(`snapshot event ${JSON.stringify(snapshotData)} can be stored as session_replay_event`, () => {
+            const data = createSessionReplayEvent(
+                'some-id',
+                12345,
+                '5AzhubH8uMghFHxXq0phfs14JOjH6SA2Ftr1dzXj7U4',
+                'abcf-efg',
+                snapshotData.events_summary,
+                snapshotSource || null
+            )
+
+            const expectedEvent: SummarizedSessionRecordingEvent = {
+                distinct_id: '5AzhubH8uMghFHxXq0phfs14JOjH6SA2Ftr1dzXj7U4',
+                session_id: 'abcf-efg',
+                team_id: 2,
+                uuid: 'some-id',
+                ...expected,
+            }
+            expect(data).toEqual(expectedEvent)
+        })
+    })
+
+    test(`snapshot event with no event summary is ignored`, () => {
+        expect(() => {
+            createSessionReplayEvent(
+                'some-id',
+                12345,
+                '5AzhubH8uMghFHxXq0phfs14JOjH6SA2Ftr1dzXj7U4',
+                'abcf-efg',
+                [],
+                null
+            )
+        }).toThrowError()
+    })
+
+    test(`snapshot event with no event summary timestamps is ignored`, () => {
+        expect(() => {
+            createSessionReplayEvent(
+                'some-id',
+                12345,
+                '5AzhubH8uMghFHxXq0phfs14JOjH6SA2Ftr1dzXj7U4',
+                'abcf-efg',
+                [
+                    {
+                        type: 5,
+                        data: {
+                            payload: {
+                                // doesn't match because href is nested in payload
+                                href: 'http://127.0.0.1:8000/home',
+                            },
+                        },
+                    },
+                    {
+                        type: 4,
+                        data: {
+                            href: 'http://127.0.0.1:8000/second/url',
+                        },
+                    },
+                ] as any[],
+                null
+            )
+        }).toThrowError()
+    })
+
+    test.each([
+        { events: [], expectedTimestamps: [] },
+        { events: [{ without: 'timestamp property' } as unknown as RRWebEvent], expectedTimestamps: [] },
+        { events: [{ timestamp: undefined } as unknown as RRWebEvent], expectedTimestamps: [] },
+        { events: [{ timestamp: null } as unknown as RRWebEvent], expectedTimestamps: [] },
+        { events: [{ timestamp: 'what about a string' } as unknown as RRWebEvent], expectedTimestamps: [] },
+        // we have seen negative timestamps from clients 🙈
+        { events: [{ timestamp: -1 } as unknown as RRWebEvent], expectedTimestamps: [] },
+        { events: [{ timestamp: 0 } as unknown as RRWebEvent], expectedTimestamps: [] },
+        { events: [{ timestamp: 1 } as unknown as RRWebEvent], expectedTimestamps: ['1970-01-01 00:00:00.001'] },
+    ])('timestamps from rrweb events', ({ events, expectedTimestamps }) => {
+        expect(getTimestampsFrom(events)).toEqual(expectedTimestamps)
+    })
+
+    function consoleMessageFor(payload: any[]) {
+        return {
+            timestamp: 1682449093469,
+            type: 6,
+            data: {
+                plugin: 'rrweb/console@1',
+                payload: {
+                    level: 'info',
+                    payload: payload,
+                },
+            },
+        }
+    }
+
+    test.each([
+        {
+            payload: ['the message', 'more strings', '', null, false, 0, { blah: 'wat' }],
+            expectedMessage: 'the message more strings',
+        },
+        {
+            // lone surrogate pairs are replaced with the "unknown" character
+            payload: ['\\\\\\",\\\\\\"emoji_flag\\\\\\":\\\\\\"\ud83c...[truncated]'],
+            expectedMessage: '\\\\\\",\\\\\\"emoji_flag\\\\\\":\\\\\\"\ufffd...[truncated]',
+        },
+        {
+            // sometimes the strings are wrapped in quotes...
+            payload: ['"test"'],
+            expectedMessage: '"test"',
+        },
+        {
+            // let's not accept arbitrary length content
+            payload: [new Array(3001).join('a')],
+            expectedMessage: new Array(3000).join('a'),
+        },
+    ])('simple console log processing', ({ payload, expectedMessage }) => {
+        const consoleLogEntries = gatherConsoleLogEvents(12345, 'session_id', [
+            consoleMessageFor(payload),
+            // see https://posthog.sentry.io/issues/4525043303
+            // null events always ignored
+            null as unknown as RRWebEvent,
+        ])
+        expect(consoleLogEntries).toEqual([
+            {
+                team_id: 12345,
+                log_level: 'info',
+                log_source: 'session_replay',
+                log_source_id: 'session_id',
+                instance_id: null,
+                timestamp: castTimestampToClickhouseFormat(
+                    DateTime.fromMillis(1682449093469),
+                    TimestampFormat.ClickHouse
+                ),
+                message: expectedMessage,
+            } satisfies ConsoleLogEntry,
+        ])
+    })
+
+    test('performance event stored as performance_event', () => {
+        const data = createPerformanceEvent('some-id', 12345, '5AzhubH8uMghFHxXq0phfs14JOjH6SA2Ftr1dzXj7U4', {
+            // Taken from a real event from the JS
+            '0': 'resource',
+            '1': 1671723295836,
+            '2': 'http://localhost:8000/api/projects/1/session_recordings',
+            '3': 10737.89999999106,
+            '4': 0,
+            '5': 0,
+            '6': 0,
+            '7': 10737.89999999106,
+            '8': 10737.89999999106,
+            '9': 10737.89999999106,
+            '10': 10737.89999999106,
+            '11': 0,
+            '12': 10737.89999999106,
+            '13': 10745.09999999404,
+            '14': 11121.70000000298,
+            '15': 11122.20000000298,
+            '16': 73374,
+            '17': 1767,
+            '18': 'fetch',
+            '19': 'http/1.1',
+            '20': 'non-blocking',
+            '22': 2067,
+            '39': 384.30000001192093,
+            '40': 1671723306573,
+            token: 'phc_234',
+            $session_id: '1853a793ad26c1-0eea05631cbeff-17525635-384000-1853a793ad31dd2',
+            $window_id: '1853a793ad424a5-017f7473b057f1-17525635-384000-1853a793ad524dc',
+            distinct_id: '5AzhubH8uMghFHxXq0phfs14JOjH6SA2Ftr1dzXj7U4',
+            $current_url: 'http://localhost:8000/recordings/recent',
+        })
+
+        expect(data).toEqual({
+            connect_end: 10737.89999999106,
+            connect_start: 10737.89999999106,
+            current_url: 'http://localhost:8000/recordings/recent',
+            decoded_body_size: 73374,
+            distinct_id: '5AzhubH8uMghFHxXq0phfs14JOjH6SA2Ftr1dzXj7U4',
+            domain_lookup_end: 10737.89999999106,
+            domain_lookup_start: 10737.89999999106,
+            duration: 384.30000001192093,
+            encoded_body_size: 1767,
+            entry_type: 'resource',
+            fetch_start: 10737.89999999106,
+            initiator_type: 'fetch',
+            name: 'http://localhost:8000/api/projects/1/session_recordings',
+            next_hop_protocol: 'http/1.1',
+            pageview_id: undefined,
+            redirect_end: 0,
+            redirect_start: 0,
+            render_blocking_status: 'non-blocking',
+            request_start: 10745.09999999404,
+            response_end: 11122.20000000298,
+            response_start: 11121.70000000298,
+            secure_connection_start: 0,
+            session_id: '1853a793ad26c1-0eea05631cbeff-17525635-384000-1853a793ad31dd2',
+            start_time: 10737.89999999106,
+            team_id: 2,
+            time_origin: 1671723295836,
+            timestamp: 1671723306573,
+            transfer_size: 2067,
+            uuid: 'some-id',
+            window_id: '1853a793ad424a5-017f7473b057f1-17525635-384000-1853a793ad524dc',
+            worker_start: 0,
+        })
+    })
+})
